feat: make TreeCleaner its own phase
