+++ main/src/ca/uwaterloo/flix/language/phase/Parser2.scala
@@ -27,123 +27,6 @@ import scala.annotation.tailrec
 
 // TODO: Add change set support
 
-// TODO: Add a way to transform Tree into a weededAst
-
-
-object TreeTransformer {
-
-  import WeededAst._
-
-  private class State(val parserInput: ParserInput, val src: Ast.Source) {}
-
-  def transform(tree: Tree, parserInput: ParserInput, src: Ast.Source): Validation[CompilationUnit, CompilationMessage] = {
-    implicit val s: State = new State(parserInput, src)
-    mapN(location(tree), visitUsesAndImports(tree), visitDeclarations(tree)) {
-      case (loc, usesAndImports, declarations) => CompilationUnit(usesAndImports, declarations, loc)
-    }
-  }
-
-  private def visitUsesAndImports(tree: Tree)(implicit s: State): Validation[List[UseOrImport], CompilationMessage] = {
-    assert(tree.kind == TreeKind.Source)
-    List.empty.toSuccess
-  }
-
-  private def visitDeclarations(tree: Tree)(implicit s: State): Validation[List[Declaration], CompilationMessage] = {
-    assert(tree.kind == TreeKind.Source)
-    List.empty.toSuccess
-  }
-
-  private def visitDefinition(tree: Tree)(implicit s: State): Validation[Declaration, CompilationMessage] = {
-    assert(tree.kind == TreeKind.Def)
-    mapN(
-      location(tree),
-      visitDocumentation(tree),
-      visitAnnotations(tree),
-      visitModifiers(tree),
-      visitNameIdent(tree.children(1)),
-      visitTypeParameters(tree),
-      visitParameters(tree),
-      visitExpression(tree),
-      visitType(tree),
-      visitTypeConstraints(tree),
-      visitConstraints(tree),
-      visitEffect(tree)
-    ) {
-      case (loc, doc, annotations, modifiers, ident, tparams, fparams, exp, ttype, tconstrs, constrs, eff) =>
-        Declaration.Def(doc, annotations, modifiers, ident, tparams, fparams, exp, ttype, eff, tconstrs, constrs, loc)
-    }
-  }
-
-  private def visitNameIdent(c: Child)(implicit s: State): Validation[Name.Ident, CompilationMessage] = {
-    Validation.Failure(LazyList.empty)
-  }
-
-  private def visitDocumentation(tree: Tree)(implicit s: State): Validation[Ast.Doc, CompilationMessage] = {
-    Ast.Doc(List.empty, SourceLocation.Unknown).toSuccess
-  }
-
-  private def visitAnnotations(tree: Tree)(implicit s: State): Validation[Ast.Annotations, CompilationMessage] = {
-    Ast.Annotations(List.empty).toSuccess
-  }
-
-  private def visitModifiers(tree: Tree)(implicit s: State): Validation[Ast.Modifiers, CompilationMessage] = {
-    Ast.Modifiers(List.empty).toSuccess
-  }
-
-  private def visitTypeParameters(tree: Tree)(implicit s: State): Validation[KindedTypeParams, CompilationMessage] = {
-    TypeParams.Elided.toSuccess
-  }
-
-  private def visitParameters(tree: Tree)(implicit s: State): Validation[List[FormalParam], CompilationMessage] = {
-    List.empty.toSuccess
-  }
-
-  private def visitTypeConstraints(tree: Tree)(implicit s: State): Validation[List[TypeConstraint], CompilationMessage] = {
-    List.empty.toSuccess
-  }
-
-  private def visitConstraints(tree: Tree)(implicit s: State): Validation[List[EqualityConstraint], CompilationMessage] = {
-    List.empty.toSuccess
-  }
-
-  private def visitExpression(tree: Tree)(implicit s: State): Validation[Expr, CompilationMessage] = {
-    mapN(location(tree))(loc => Expr.Error(Parse2Error.DevErr(loc, "expected expression")))
-  }
-
-  private def visitType(tree: Tree)(implicit s: State): Validation[Type, CompilationMessage] = {
-    mapN(location(tree))(Type.Unit)
-  }
-
-  private def visitEffect(tree: Tree)(implicit s: State): Validation[Option[Type], CompilationMessage] = {
-    None.toSuccess
-  }
-
-  private def location(tree: Tree)(implicit s: State): Validation[SourceLocation, CompilationMessage] = {
-    mapN(treeBegin(tree), treeEnd(tree)) {
-      case ((b_line, b_col), (e_line, e_col)) =>
-        SourceLocation(Some(s.parserInput), s.src, SourceKind.Real, b_line, b_col, e_line, e_col)
-    }
-  }
-
-  @tailrec
-  private def treeBegin(tree: Tree): Validation[(Int, Int), CompilationMessage] = {
-    tree.children.headOption match {
-      case Some(Child.Token(token)) => (token.line, token.col).toSuccess
-      case Some(Child.Tree(tree)) => treeBegin(tree)
-      case _ => Validation.Failure(LazyList.empty)
-    }
-  }
-
-  @tailrec
-  private def treeEnd(tree: Tree): Validation[(Int, Int), CompilationMessage] = {
-    tree.children.lastOption match {
-      case Some(Child.Token(token)) => (token.line, token.col).toSuccess
-      case Some(Child.Tree(tree)) => treeEnd(tree)
-      case _ => Validation.Failure(LazyList.empty)
-    }
-  }
-}
-
 /**
  * Errors reside both within the produced `Tree` but are also kept in an array in `state.errors`
  * to make it easy to return them as a `CompilationMessage` after parsing.
@@ -180,7 +63,7 @@ object Parser2 {
     case class Closed(index: Int) extends Mark
   }
 
-  def run(readRoot: ReadAst.Root, root: Map[Ast.Source, Array[Token]], entryPoint: Option[Symbol.DefnSym])(implicit flix: Flix): Validation[WeededAst.Root, CompilationMessage] = {
+  def run(root: Map[Ast.Source, Array[Token]])(implicit flix: Flix): Validation[Map[Ast.Source, Tree], CompilationMessage] = {
     if (flix.options.xparser) {
       // New lexer and parser disabled. Return immediately.
       return Validation.Failure(LazyList.empty)
@@ -189,27 +72,14 @@ object Parser2 {
     flix.phase("Parser2") {
       // Parse each source file in parallel and join them into a WeededAst.Root
       val results = ParOps.parMap(root) {
-        case (src, tokens) => mapN(parseWithFallback(src, tokens))(trees => src -> trees)
+        case (src, tokens) => mapN(parse(src, tokens))(trees => src -> trees)
       }
 
-      mapN(sequence(results))(_.toMap).map(m => WeededAst.Root(m, entryPoint, readRoot.names))
+      mapN(sequence(results))(_.toMap)
     }
   }
 
-  private def parseWithFallback(src: Ast.Source, ts: Array[Token])(implicit flix: Flix): Validation[WeededAst.CompilationUnit, CompilationMessage] = {
-    parse(src, ts) match {
-      case s: Success[_, _] => s
-      case _ =>
-        println(s"${src.name}\tFalling back on Parser")
-        Parser.parseRoot(src) match {
-          case Success((_, unit)) =>
-            Weeder.visitCompilationUnit(src, unit).map(_._2)
-          case _ => Validation.Failure(LazyList.empty)
-        }
-    }
-  }
-
-  private def parse(src: Ast.Source, ts: Array[Token]): Validation[WeededAst.CompilationUnit, CompilationMessage] = {
+  private def parse(src: Ast.Source, ts: Array[Token]): Validation[Tree, CompilationMessage] = {
     implicit val s: State = new State(ts, src)
     source()
     val tree = buildTree()
@@ -219,7 +89,7 @@ object Parser2 {
     if (s.errors.length > 0) {
       Validation.Failure(LazyList.from(s.errors))
     } else {
-      TreeTransformer.transform(tree, s.parserInput, s.src)
+      tree.toSuccess
     }
   }
